requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
asttokens==3.0.0
attrs==25.1.0
chumpy==0.70
decorator==5.2.1
executing==2.2.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.3.0
grpcio==1.70.0
idna==3.10
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
lightning==2.5.0.post0
lightning-utilities==0.14.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==6.30.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pytorch-lightning==2.5.0.post0
PyYAML==6.0.2
scipy==1.15.2
setuptools==76.0.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.6.0
torchmetrics==1.6.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.18.3
